Reimplemented folding

+++ src/compiler/compiler.c
@@ -8,7 +8,7 @@ void compiler_init( compiler* compiler, heap* heap, parser* parser )
 /* Yields a base opcode for binary and unary operators. Base opcodes all
 ** start from the first existing value_type, e.g. value_i64
 */
-u32 tk_to_op( token_type tk )
+i64 tk_to_op( token_type tk )
 {
 	static const opcode ops[ tk_n ] = { OPCODES( op_, X_OP_MAP ) };
 	return ops[ tk ];
